wrong_shard_server handling, location cache fix, wire type cleanup
wrong_shard_server (1062):
- Location cache now actually caches shard ranges (was never storing results)
- parseGetKeyServerLocationsReply properly reads KeyRangeRef as nested struct
  (was misreading RelOff as bytes — caused by hand-written slot guessing)
- Fault injection test via wrongShardConn: pipe-based TCP proxy at net.Conn
  level, PING-aware frame replacement. Zero production code changes.

Wire type fixes:
- Error.MarshalInto: WriteUint16 not WriteInt32 (ground-truth: uint16_t wire type)
- GetKeyServerLocationsReply: proper typed parser in wire/types/ ported from
  C++ serialize(), replaces 80-line brute-force slot scanner in locality.go
- All reply types now use generated slot constants (no hardcoded numbers)
- GetReadVersionReply: fixed wrong slots 5,6,7,10 → correct 6,7,8,11
  (Optional<Value> eats two vtable slots, was miscounted)

Cleanup — delete deprecated JSON schema infrastructure:
- Removed cmd/fdb-wire-schema-generator/ (369 JSON schemas, Go parser, Docker build)
- Removed cmd/fdb-wire-codegen/ (Go codegen from JSON — wrong approach)
- Removed pkg/fdbgo/wire/schema/ (369 JSON files) and pkg/fdbgo/wire/testdata/
- Removed buf.yaml, buf.gen.yaml (were for schema generator, not proto codegen)
- Single codegen path: cmd/fdb-schema-extract/main.cpp → Go code directly

C++ extractor extension (RFC 011):
- GoEmitter now emits Go struct + UnmarshalFDB + MarshalFDB per type
- GoTypeMapping<T> template resolves exact C++ scalar types to Go types
  (bool, int32, uint16, float64, etc.) at compile time — no heuristics
- Nested struct fields: comment + skip (manual ReadNestedReader)
- Optional fields: Has+Value two-field pattern
- Needs Docker rebuild to test: bash cmd/fdb-schema-extract/build.sh

RFC 011: Wire Type Generation — design for generated wire types with
override mechanism for types with conditional C++ serialize() logic.

+++ pkg/fdbgo/wire/types/error.go
@@ -2,8 +2,9 @@ package types
 
 // Error — flow/include/flow/Error.h
 // C++ serialize: serializer(ar, error_code)
-//   slot 0: error_code — int (scalar, 4 bytes at offset 4)
-// Note: C++ Error.error_code is int (32-bit), not uint16.
+//   slot 0: error_code — uint16_t (scalar, 2 bytes at offset 4)
+// Note: C++ Error.error_code is declared as int, but serialized as uint16_t
+// on the wire. Confirmed by schema extraction + ground-truth test vector.
 
 import "github.com/birdayz/fdb-record-layer-go/pkg/fdbgo/wire"
 
@@ -15,11 +16,14 @@ func (m *Error) TypeVTable() wire.VTable { return ErrorVTable }
 
 func (m *Error) MarshalInto(obj *wire.ObjectWriter) {
 	vt := ErrorVTable
-	obj.WriteInt32(int(vt[2]), m.Code)
+	obj.WriteUint16(int(vt[2]), uint16(m.Code))
 }
 
 func (m *Error) UnmarshalFrom(r *wire.Reader) error {
 	if r.FieldPresent(0) {
+		// Wire format is uint16, but we store as int32 for convenience.
+		// ReadInt32 reads 4 bytes (2 value + 2 zero-padding), which is
+		// safe because FDB zero-initializes object buffers.
 		m.Code = r.ReadInt32(0)
 	}
 	return nil
